backend/requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
neo4j==5.14.0

# Document Processing
PyPDF2==3.0.1
python-pptx==0.6.21
python-docx==0.8.11

# NLP & ML (lightweight versions)
spacy==3.7.2
dateparser==1.1.8
numpy==1.26.2
transformers==4.35.0
torch==2.1.0

# Authentication & Security
PyJWT==2.11.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
